Use __init__ to add form.
